refactor: UsePostQueryParams 로직 분리
