+++ src/feature/postManagement/model/index.ts
@@ -1,1 +1,2 @@
 export * from "./usePostManagement"
+export * from "./usePostQueryParams"
